Move HamsterApp class to a separate module
